+++ docs/source/small.rst
@@ -1,0 +1,8 @@
+Small Molecules
+===============
+
+
+.. toctree::
+    :maxdepth: 1
+
+    NCore <_autosummary/stk.small.NCore>
