+++ src/stk/_internal/functional_groups/functional_group.py
@@ -83,7 +83,7 @@ class FunctionalGroup:
     actually get made during construction. Finally, you will pass an
     instance of your :class:`.ReactionFactory` subclass to the
     chosen :class:`.TopologyGraph` you want to make, for example
-    :class:`~.polymer.linear.linear.Linear`, and your custom
+    :class:`~.polymer.linear.Linear`, and your custom
     modification will take place.
 
     See Also:
