--- src/stk/_internal/topology_graphs/polymer/linear/__init__.py
@@ -1,1 +1,0 @@
-
