+++ src/stk/_internal/topology_graphs/small/__init__.py
@@ -1,0 +1,10 @@
+"""
+Small Molecules
+===============
+
+#. :class:`.NCore`
+
+"""
+
+from . import vertices  # noqa
+from .ncore import *  # noqa
