+++ src/stk/small.py
@@ -1,0 +1,11 @@
+from stk._internal.topology_graphs.small.ncore import NCore
+from stk._internal.topology_graphs.small.vertices import (
+    CoreVertex,
+    SubstituentVertex,
+)
+
+__all__ = [
+    "NCore",
+    "CoreVertex",
+    "SubstituentVertex",
+]
