+++ tests/molecular/topology_graphs/vertex/cases/small/test_get_position.py
@@ -1,0 +1,1 @@
+from ..._test_get_position import *  # noqa
