+++ tests/molecular/topology_graphs/vertex/cases/small/test_with_position.py
@@ -1,0 +1,1 @@
+from ..._test_with_position import *  # noqa
